Added VRCamera Class
Initial version of VRCamera. Currently the class keeps track of 3
cameras, a left, right and a center camera. Center is for standard
camera usage without HMD and for future Mono-stereo impl.

Currently missing all VR functionality in order to keep track off and
update the positions of the cameras through the HMD.

Issue: #6