Added submit and sync calls for VR
This commit adds two submit calls for the VR namespace that take a
ColorBuffer with a texture array or two a left and right eye ColorBuffer and submits
them to the screen.

Also each function that uses the HMD now checks if one is present and
prints a debug message and returns if none is.

Issue: #11