Add VR Submit and sync to GraphicsCore
This commit adds a function to submit the ColorBuffer in the graphics
core to the HMD and adds the Sync function before the present. Sync
function should be tested and moved if necessary. Also sumbission should
be tested with regards to using the correct ColorBuffer and that
Texture2DArray works as intended.

Issue: #11