Fix SSAO, refactor + warning fix
SSAO now properly uses sub-SRVs when rendering, fixing the graphical
bugs it had. This required a small refactoring of the SSAO code, as it
now needed to take the camera as a parameter.

We also noticed a erroneous use of 'sqrt' instead of 'sqrtf' causing
some warnings. We fixed this.