Fix raster scene transform; Re-add default texture
The global scene transform was being applied incorrectly: model *
viewproj instead of the correct viewproj * model. This is now fixed, so
scenes with scene-global transforms, such as the bistro scene, are also
rendered correctly in rasterization.

The removal of default textures made the bistro scene crash, since there
are materials missing a specular texture. This commit reintroduces the
g_DefaultTexture, which is set to all zero.