Fix compile error, remove anyhit shader
This commit fixes the compile error introduced by last commit
"~checkpoint". The problem was the include order was incorrect. Instead
of fixing the include order, PipelineState was included in Camera to
completely illeviate the issue.

This was done last commit, but this is the first compiling commit with
the deletion of the unused anyhit shader.

Issue #61