[~]Back to LowResPass
This commit goes back to using the LowResPass buffer as it needs to
be full resolution and not simply a mip. Therefore this commit goes
over to using the mip for the final LowRes image which was also the
original plan.

Currently has some graphical issues with low res not showing up.