Save DB clearval, add depth bounds, fix CS depth
The depth buffer clear value was not being saved during creation. This
is now saved to the appropriately and already used member variable
m_clearValue in DepthBuffer.

In order to help debug Screen Space Reflections, changing the depth
range is useful. This feature was missing from the GraphicsContext,
so I've added a SetDepthBounds(float min, float max) to alleviate the
issue.

Lastly, I've removed the linearization of the depth buffer from
the HybridSSR compute shader.