[~]Complete rework of buffer structure
This commit removes all low res buffers in favor of single buffers with
mip levels to represent lower resolution versions. Currently only the high
res version of these are in use and the program produces a black screen.
No errors or warnings are present and the program does indeed run.

Next is working out the issues with only high res and then implement
systems to alternate between mips where possible.