Adjusted motion detection
It was being done using the value of the previous high res frame which is the altered high resolution frame. It is now being done using the low passed frame. This is still not entirely correct as it is also downsampled and upsampled in the paper, but it should be a better approximation than before.