Fix plant border black in bistro bug
The IsReflection parameter of HitShaderConstants was always being set to
false. This meant that in the DiffuseHitShader the 'dampen' value was
being incorrectly calculated.

This does not really affect performance at all, since the recursive ray
tracing is performed based on material reflectivity exclusively. Redo of
profiling is not required.

This commit adds a 'refl' parameter to the RaytraceDiffuse call, so the
IsRelfection can be controlled for holefilling.