Finalize normal buffer reorganizing
This commit finalizes the transition from
        [normalX, normalY, normalZ, specular]
to
        [normalX, normalY, ratio, specular]

This change was made to fix a visual artefact when lerping specular to
use as reflectivity ratio. Lerping the specular caused the lighting to
be wrong in the blending range, which is very visible. The ratio is now
a dedicated parameter, which is not used for anything else.

There is still the following issues:
  - Normal Texture Strength = 0 causes reflections to disappear.
  - This commit includes 3 normalizations that were not there before.
    Check if these are necessary.