Add normal flipping option
Some scenes, specifically the bistro scenes, need their normals inverted
in the shaders. This commit adds a parameter to the PSConstants and
HitShaderConstants structures to allow for this invertion.

The flip is implemented as a numerical multipler in the scene data. If
the normal needs to be "FlipNormal" value is set to -1, 1 otherwise.